Implemented methods related to guessing. A few other changes.
Added actions to take as soon as a player guesses a letter.
Added actions to take whenever a player makes a correct and wrong guess.
Changed all guessing letters to lowercase for convenience
When a word is < 9 characters long, rather than display "_" in the character's place, just display empty space.